server/requirements.txt
astroid==2.3.2
atomicwrites==1.3.0
attrs==19.3.0
blinker==1.4
Click==7.0
colorama==0.4.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
get==2019.4.13
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.2
Pillow==6.2.1
pluggy==0.13.0
post==2019.4.13
public==2019.4.13
py==1.8.0
PyJWT==1.4.2
pylint==2.4.3
pymongo==3.9.0
pyparsing==2.4.2
pytest==5.2.2
query-string==2019.4.13
request==2019.4.13
six==1.12.0
typed-ast==1.4.0
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
